Loosen checks for functions expected to take a BiConsumer
Fixes packets in 22w18a. We used to explicitly check that args[1] was a java/util/function/BiConsumer but in 22w18a Mojang added their own subclass of that interface (nested in packetbuffer, and with a defaulted method called asOptional) which breaks that explicit check.  We could add an additional check to see if it is that custom interface, but that probably is not needed (it'll blow up if the wrong type is used with or without the check, just in a different way).