Fix entitymetadata for 22w18a
This is where that asOptional function is used.  Which means more work to make lambdas behave more accurately, since now we have a case where one lambda calls another lambda, and the existing code didn't handle that correctly.  The new code isn't 100% perfect either, but it works in the current cases at least.